+++ app/utils/get-projection-by-name.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry/utils/get-projection-by-name';