+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* `flexberry-groupedit` component:
+    * Fix ordered property computing when `modelProjection` are string.
 
 ## [2.1.0-beta.4] - 2019-04-15
 ### Added